Change BUILD.md to HACKING.md, document local repos, like enmasse
enmasse have a HACKING.md document to explain how to do
development-type things on the project. Really our BUILD.md
is a similar thing. I've expanded it to include basic instructions
for building the running images with a local docker repo.

I've also fixed the long lines, so it's nicer to read in a terminal.