An interface to cope with partition assignment, but I'm not sure this is the best approach
One problem with it is that it assumes we can ask for an assignment separately
from actually increasing partitions/replication factor. That might not fit with
how the cluster balancer actually works.

Also it doesn't cope at all with another aspect of data-movement: Throttling.

So a better way might be to get rid of PartitionAssignment and have subclasses
or parameterised KafkaImpl which worries about assignment.