* Rename the CreateClusterOperation -> SimpleClusterOperation, since it will   suffice for deletion as well * Delete (Kafka|KafkaConnect|Zookeeper)ClusterOperation classes, because   they're not doing anything useful * Refactor Delete(Kafka|KafkaConnect|Zookeeper)ClusterOperation to   subclass SimpleClusterOperation
Note: I think there was a bug before around how the volume claims were
deleted. Before we would complete the overall `handler` as a result of the
join() on futurePersistentVolumeClaims. This means that the handler was
already completed irrespective of the completion of the statefulset and
service futures. We now add the PVC deletion futures to the overall result,
which means the handler will only be completed once, though with some very minor
loss of logging info (we don't know specifically it was the PVC deletions
which failed).