Better rolling (#462)
Fixes #363 by changing the way StatefulSets are updated/rolled so that if the CO crashes part way through it will complete the roll when it restarts.

Using the same logic as before, whenever we patch in a way that needs a rolling update we update the SS template, incrementing the value of a 'generation' annotation. This happens as part of the SS patch. Previously, we only did the rolling update if the patch operation indicated that it was needed, but now (because a rolling update may be necessary even if the last patch was a noop) we always call maybeRollingUpdate(), and let that method decide which pods to roll. Since rolled pods will pick up the new generation number from the SS's template we can distinguish rolled pods from unrolled ones. Thus we only actually rolls pods whose generation annotation does not match the StatefulSets, and so maybeRollingUpdate() will be a noop iff all the pods' generation numbers match.