Wait for pod readiness in maybeRestartPod()
Only complete the future returned from maybeRestartPod() when the pod is ready
whether or not it was actually deleted. This will avoid a possible loss of
cluster availability when, over successive reconciliations, a
pod is deleted, but not successfully restarted by the infrastructure.
By not checking for readiness each reconciliation would take down the next pod,
until, possibly, all pods were down.

In the normal case and when the pod is not restarted this will be cheap.

There are three production call sites:

1. The one in ZookeeperSetOperator was already composing with a readiness
   check, so that is be removed and functionality is unchanged.
2. The one in StatefulSetOperator.maybeRollingUpdate() will now wait,
   avoiding the problem described above.
3. The one in StatefulSetOperator.maybeDeletePodAndPvc will now wait,
   avoiding the problem described above.