Factor out ZookeeperLeaderFinder, use BackOff and vertx net client
To make the zookeeper rolling algortihm more robust we now use a
retry-with-backoff strategy when finding the zookeeper leader.
In the event that we cannot find the leader we proceed with the
reconciliation anyway (it could be the zookeeper cluster is in a broken state
and needs a restart to resolve it).

Move and reuse the existing BackOff class from the Topic Operator.

Use vertx for the networking, which simplifies certificate and key handling.