Factor out ZookeeperLeaderFinder, use BackOff and vertx net client (#1303)
To make the zookeeper rolling algorithm more robust when finding the zookeeper leader we now use a
retry-with-backoff strategy. In the event that we cannot find the leader we proceed with the reconciliation anyway (it could be the zookeeper cluster is in a broken state and needs a restart to resolve it).

Also switch to using vertx for the networking, which simplifies certificate and key handling.