Add test for and fix #1221 and #472
Use the private topic store as a mapping between the topic name in Kafka and
the KafkaTopic metadata.name. Previously we were incorrectly just using the
Kafka name assuming it must come from a KafkaTopic with that metadata.name.

This complicates the reconcileAll algorithm, because we cope with the
possibility that the private topic does not exist, so we have to pass that
undetermined state through to where we have to reconcile the KafkaTopics
present in kube.