Support for more informative deprecation in docs, logs (#1462)
Fixes #1460 by adding a new annotation for deprecating properties in CRs which is consumed by the DocGenerator and by the validation at runtime. Unfortunately a new module is necessary because (unlike existing annotations) the annotation's .class needs to be on both the doc-generation-time and runtime classpaths, but this seemed simpler than the alternatives.