Reads should be included in scope of lock
During initial and timed reconciliations we were only taking the topic lock
after some of the remote resources had been fetched.
This meant there was a race between the timed reconciliation and any topics
touched immediately after the TO was started.
This meant it was possible for the TO to get into a state where
reconciliations of a given topic proceeded endlessly.

The topic lock is now acquired before reading *any* of the remote resources
required for a reconciliation.