Ignored added events during initial reconciliation
When a watcher is created is receives added events for all existing matching resources.
This means there's a race between the watcher and the initial reconciliation.
In some circumstances this race results in incorrect reconciliation.
For example if a topic in Kafka was deleted while the TO was down and the added event occurs before the initial reconciliation then the topic can get re-added.