Kube-side changes to #replicas should result in NotReady status
This changes the Topic Operator's handling of Kube-side changes to
KafkaTopic.spec.replicas by ensuring the reconciliation results in a
NotReady status. If the replication factor is changed on the Kafka side
then the reconciliation updates KafkaTopic.spec.replicas and results in
Ready status.

Note that, because the TO doesn't watch the relevent ZK nodes, the
KafkaTopic.spec.replicas only gets updated during a periodic reconciliation.

An integration test is added for the revised semantics.

Because change in replicas has never been supported I have also removed
unused the code which notionally allowed for the TO to perform partition
reassignment, but in fact was never used.

Fix #1847, fix #691

Signed-off-by: Tom Bentley <tbentley@redhat.com>