Refactor MockKube and fix a problem with pod recreation.
The problem was due to the fact that we were using watchers
to recreated delected pods in a statefulset. Because the "recreation watcher"
was fired first (and because the whole thing is single threaded) it meant
we recreated the pod (for fired the creation event) before other watchers had
been told of the deletion.

The solution was to introduce a separate Observer mechanism so that such
recreations can be done after all watches are fired. This should be easily
extensible for other things should that be necessary.

Signed-off-by: Tom Bentley <tbentley@redhat.com>