feat: Enable internal connections for JMX port.
Created a new field in the Kafka CRD which would allow you to enable the
JMX metrics on the kafka broker. When the field is specified, the kafka
broker starts up with port 9404 for the JMX. Then it exposes 9404 on the
kafka headless service to be able to collect the metrics.

Signed-off-by: Julian Goh <juliangohsl@gmail.com>