feat: Enable internal connections for JMX port.
Created a new field in the Kafka CRD which would allow you to enable the
JMX metrics on the kafka broker. When the field is specified, the kafka
broker starts up with port 9999 for the JMX. Then it exposes 9999 on the
kafka headless service to be able to collect the metrics.

Enabled the ability to start up port with password and username
protection. Renamed Metrics to prometheus metrics to avoid confusion

Signed-off-by: Julian Goh <julian.goh@ibm.com>