Feat: enable internal JMX Connections on a secure or insecure JMX port on Kafka Brokers (#2268)
* feat: Enable internal connections for JMX port.

Created a new field in the Kafka CRD which would allow you to enable the
JMX metrics on the kafka broker. When the field is specified, the kafka
broker starts up with port 9999 for the JMX. Then it exposes 9999 on the
kafka headless service to be able to collect the metrics.

Enabled the ability to start up port with password and username
protection. Renamed Metrics to prometheus metrics to avoid confusion

Signed-off-by: Julian Goh <julian.goh@ibm.com>

* fix: Address PR Comments

Renamed prometheusMetrics to metrics and moved change in
change log

Signed-off-by: Julian Goh <julian.goh@ibm.com>

* fix: Address PR Comments

Refactored name and removed unnecessary checks.

Signed-off-by: Julian Goh <julian.goh@ibm.com>

* fix: address PR comments

added new field to expand KafkaJmxAuthentication

Signed-off-by: Julian Goh <julian.goh@ibm.com>

* fix: Fix forever recreating secret

Secret would always get generated with new data so it would Kube would
update the secret which means that future pods will use an invalid
username. Changed it so that the cluster operator would do nothing if
the secret exists.

Refactored test to not iterate through additional param.

Added a more generic way to add different jmx authentication
methods

Signed-off-by: Julian Goh <julian.goh@ibm.com>

* refactor: Address PR comments

rename fields and classes. Changed docs to make more sense.

Signed-off-by: Julian Goh <julian.goh@ibm.com>