feat: Fail make docker_build if yq version not 3
Multiple people myself included have been bitten by having
the wrong yq version. The docs have been updated to reflect
this, but considering the builds will fail with obscure
mistemplated file errors it would be better to check upfront
the yq version.
If a user does not have yq version 3.*.* then it will fail
the build and print the warning in red.

Signed-off-by: Samuel Hawker <samuel.hawker@ibm.com>