feat: Add warnings to CR status conditions
This commit extends the use of the conditions list in the CR status
to include warnings about potential problems and recommendations for
best practice. It introduces a single point in the reconcile loop
where the cluster spec can be checked for things that the user
might not be aware of.

The list of warnings that can be generated is relatively short, but
the intent is to introduce the scaffolding where new checks and
warnings could later be introduced.

Contributes to: #2609

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>