feat: Enable structural schemas in Strimzi CRDs
Strimzi relies on being able to store some fields not explicitly
defined in the CRD (such as in metrics or config objects).

The default behaviour for this is changing in Kubernetes, so ahead of
this change, this commit makes it explicit exactly where storing
unknown fields is required (rather than depending on it across the
entire CRD).

This commit updates CrdGenerator to set preserveUnknownFields to false
for the CRD as a whole, and uses x-kubernetes-preserve-unknown-fields
to override this for objects where pruning would cause a problem.

Signed-off-by: Dale Lane <dale.lane@uk.ibm.com>