feat: Unambiguate deletion command
In a cluster with multiple custom resource kinds with shortname
kafkatopics, the test can fail to delete the topic. By using the
CRD name, disambiguates the type for deletion.

Also added a log line to the MetricsST to signal its start,
multiple runs locally have seen it hang with no printed output
so this output makes the runner aware of when the test actually begins

Signed-off-by: Samuel Hawker <samuel.hawker@ibm.com>