Add explicit existaece check on Kafka.topicMetadata()
Since a null result from topicMetadata() implies non-existence
is was prone to a race where we queried a non-controller broker for the
metadata if that broker has not yet processed an UPDATE_METADATA request
from the controller (e.g. the topic was recently created or deleted). By
adding an explicit check for existing which will be handled by the controller
we should have somewhat more consistent responses.

Signed-off-by: Tom Bentley <tbentley@redhat.com>