Topic operator improvements (#3982)
Try to fix #1157 and possibly other issues where the TO behaves incorrectly wrt creation or deletion.

When we observe the deletion of a topic via ZK don't assume the topic deletion has propagated to observations of the topic via the Admin client. Instead wait for the Admin client to confirm that the topic doesn't exist before proceeding with the reconciliation. Using a null result from `Kafka.topicMetadata()` to imply non-existence
was prone to a race where we queried a non-controller broker for the
metadata and that broker had not yet processed an UPDATE_METADATA request
from the controller (e.g. the topic was recently created or deleted). By
adding an explicit check for existence (handled by the controller because it's a `createTopics(validateOnly=true)`)
we should have more consistent responses.

Signed-off-by: Tom Bentley <tbentley@redhat.com>