feat: Add annotations to perform connector/task restart operations
 - This commit implements the first part of the 'Restarting Kafka
Connect connectors and tasks' proposal
(https://github.com/strimzi/proposals), adding new annotations that
cause the operator to restart connectors or tasks. The annotations can
be applied to the KafkaConnector, and the KafkaMirrorMaker2 custom
resources. The annotation acts as a trigger for a single restart call
by the operator, and is removed from the CR when the restart REST API
call is successfully called.

Signed-off-by: Andrew Borley <borley@uk.ibm.com>