Add client rack-awareness support to Strimzi Bridge pods
If you configure a rack topologyKey for a KafkaBridge CRD then an init container will
be added to the bridge pod to retrieve the value of rack.topologyKey and write it into
a file in `/opt/strimzi/init`. This `/opt/strimzi/init` volume is also mounted into the
bridge container so it can retrieve the rack.id at runtime and use it for consumer
configuration.

Why:
You could configure your Kafka Bridge to prefer to consume from a broker in the same
zone, potentially saving on inter-zone transfer costs. This may cost in overall
message latency as you may be consuming from a follower for a topic-partition.

`/opt/strimzi/init` was selected as the mount because strimzi-bridge startup scripts use
a STRIMZI_HOME instead of KAFKA_HOME, this keeps the configuration within /opt/strimzi
instead of introducting /opt/kafka to the image.

Signed-off-by: Robert Young <robeyoun@redhat.com>