ZookeeperRoller considers unready pods
This fix makes the ZookeeperRoller:
1. re-order the pods so that non-ready pods are evaluated first
2. wait for pods that do not require a restart to become ready

This means that in the a scenario with a pod stuck in a non-ready state it will
evaluate the non-ready pod first on each run and await readiness. Likely ending
in an exception.

Why:
Previously a bad pattern could occur where ZookeeperRoller restarted a pod and it
did not become ready. This eventually hit a timeout that failed the roll but
on the next periodic reconciliation that pod would not be eligible for a restart
and the roller would then restart the next node. This would eventually take the
whole Zookeeper cluster offline if, for example, the pods couldn't be scheduled.

Note:
There are further specialisations in the KafkaRoller that we could add/share with
the ZookeeperRoller, like a short-circuit for pods that are stuck in pending due
to them being unschedulable. This would save waiting for the readiness timeout.

Fixes #1001

Signed-off-by: Robert Young <robeyoun@redhat.com>