Replace sensitive information mask with "omitted"
Using * chars for sensible information mask (passwords and keys from report.sh) breaks YAML parsers.
The * char is used as part of the anchors and aliases feature in YAML. The parser raises and error
when it finds it and it is not expected. We can use quotes so that the parser sees the mask as a
string, or use a different replacement as proposed here.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>