Await readiness of the leader during the first pass
Why:
This covers an edge-case where we could determine a zookeeper pod is non-ready but then
it is determined to be the leader. In this case if we await readiness of the leader
before trying to restart the followers then we get a chance to prevent those restarts
if it is really down. It seems unlikely since we will have contacted the pod to determine
if it is leader.

Signed-off-by: Robert Young <robeyoun@redhat.com>