Improve ZooKeeperRoller logging
The restart debug message was misleading as it was logged during construction of the
future chain, not when a pod was actually restarted. There is an info log that
covers the actual restart.

The waiting for readiness log was duplicated by logging inside the readiness call

An incorrect log would have been written for leaders that require a restart saying
that they don't need restart.

Signed-off-by: Robert Young <robeyoun@redhat.com>