Fix failing tests exposed by vertx-junit5 upgrade
I ran them on main and observed that the assertions were failing silently without
causing the test to fail. There is an open vertx-junit5 issue that looks relevant

These tests failed due to an unexpected error message. It expected the pods to not
roll because they had no restart reasons but the roller determines they need to be
reconfigured due to a logging config difference.

https://github.com/eclipse-vertx/vertx-junit5/issues/106
Signed-off-by: Robert Young <robeyoun@redhat.com>