Upgrade to vertx-4.3.4
We had to change how we use vertx shared worker pools to prevent them being
automatically closed.

There was a change in vertx 4.3.0 that meant that when a SharedWorkerExecutor is
finalised it is closed, which calls close on the executors SharedWorkerPool. If there
are no references to the SharedWorkerPool it will be shutdown. So to keep the pool
alive we need to have at least one reference to an executor to keep it alive.

vertx issue: https://github.com/eclipse-vertx/vert.x/issues/4345
vertx PR: https://github.com/eclipse-vertx/vert.x/pull/4352

Signed-off-by: Robert Young <robeyoun@redhat.com>