Delete cold-backup bash script and related ST
This script is too slow in the current form because it does not support incremental backups
and it uses gzip instead of pigz (a parallel implementation of gzip). Additionally, as the
name implies, the cluster needs to be stopped for the entire duration of the backup, which
is a deal breaker for many production clusters. The script can be adapted to do online
backups, but we are not sure if this would always work fine, for example when topics are
deleted and recreated during the backup, or how it would deal with inflight transactions.
Finally, this kind of complexity is not appropriate for a bash script and there are better
tools for online storage snapshotting.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>