Fix failing tests when HTTP/HTTPS proxy env vars are set
Even if you are not behind a proxy, you may have NO_PROXY set.
For example, on Fedora 36 with Gnome the default configuration is empty proxy with NO_PROXY set.
In general, many corporate developers are behind proxies, so this would be useful for them.

Related test classes:

- KafkaConnectClusterTest
- KafkaExporterTest
- EntityTopicOperatorTest
- EntityUserOperatorTest
- KafkaBridgeClusterTest
- KafkaMirrorMaker2ClusterTest
- KafkaMirrorMakerClusterTest
- CruiseControlTest

Because of that, KafkaConnectAssemblyOperatorPodSetTest.testBuildChangedCluster was also
failing with a wrong image build hash.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>