Add provider to host the env vars shared by all containers
The environment variables shared by all containers are statically defined in ContainerUtils.
They include proxy env vars that are also set in the Dockerfile for KafkaConnect builds.

When proxy env vars are set or updated, some tests fail due to the following reasons:

The Dockerfile hash changes:

- KafkaConnectAssemblyOperatorPodSetTest
- KafkaConnectBuildAssemblyOperatorKubeTest
- KafkaConnectBuildAssemblyOperatorOpenShiftTest

The expected test result changes:

- KafkaConnectClusterTest
- KafkaExporterTest
- EntityTopicOperatorTest
- EntityUserOperatorTest
- KafkaBridgeClusterTest
- KafkaMirrorMaker2ClusterTest
- KafkaMirrorMakerClusterTest
- CruiseControlTest

This change adds the SharedEnvironmentProvider interface with a DefaultSharedEnvironmentProvider implementation.
This is propagated to all components, and can be replaced with a MockSharedEnvironmentProvider for testing.
The mock implementation makes the unit and integration test suites independent from the test environment.
Additionally, this completely replace the previous fix which only addresses the second issue.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>