Add support for pausing KafkaTopic reconciliations with UTO
This change adds support for pausing KafkaTopic reconciliations with the Unidirectional Topic Operator (UTO).
After the strimzi.io/pause-reconciliation annotation is set to "true" on a KafkaTopic, no configuration change will be reconciled.
All valid changes will be reconciled as soon as the annotation is set to "false" or removed.
Note that KafkaTopic deletions are always reconciled, even when the resource is paused.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>