Add operation metrics to the Unidirectional Topic Operator
The UTO logs some timer metrics for the different Kubernetes and Kafka operations.
They are not much useful in logs, so this change proposes to add them to the main set of metrics.
There is the possibility that the design will evolve to a batch per Kubernetes event type, but I think we will still be interested in these aggregations.

- addFinalizer: strimzi.add.finalizer.duration
- removeFinalizer: strimzi.remove.finalizer.duration
- updateStatus: strimzi.update.status.duration
- incrementalAlterConfigs: strimzi.alter.configs.duration
- createPartitions: strimzi.create.partitions.duration
- createTopics: strimzi.create.topics.duration
- describeTopics: strimzi.describe.topics.duration
- describeConfigs: strimzi.describe.configs.duration
- deleteTopics: strimzi.delete.topics.duration
- listPartitionReassignments: strimzi.list.reassignments.duration

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>