Fix UTO NPE with missing partitions and/or replicas
If you create or update a KafaTopic without partitions and/or replicas, the operator fails with a NPE. In all these cases, the operator should apply the default cluster values. Note that empty spec (spec: {}) can be reconciled, but it is different from missing spec, which should not be reconciled. We also need to match the BTO behavior here. In the future we will make the spec required at the schema level.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>