Improve UTO logs
This change improves the UTO logs. It now logs every periodic reconciliation and every batch reconciliation. It also logs individual topic reconciliations in case of failure. All YAML paylodas have been replaced with topic name.

---

The KafkaTopic informer resync configuration had a small issue which was causing the skip of some periodic reconciliations. They were often triggering at 4 minutes, instead of 2 minutes.

The informer interval acts like a heartbeat, then each handler interval will cause a resync at some interval of the overall heartbeat. The closer these values are together the more likely it is that the handler skips one informer intervals. Setting both intervals to the same value generates just enough skew that when the informer checks if the handler is ready for resync it sees that it still needs another couple of micro-seconds and skips to the next informer level resync.

This is fixed by introducing a small fixed resync period for the informer. The resync operation is all in memory and results in a noop most of the time, so this causes no harm.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>