Fix paused connector status when Connect is scaled to zero
The common strategy across operators is that a paused resource should overwrite any error in terms of status update.
This is currently not happening when a connector is paused after Connect is scaled to zero.

The connector is first updated to `NotReady` by `connectorEventHandler`, and then `ReconciliationPaused` by `reconcileConnectors`.
The `testConnectorResourceMetricsScaledToZero` fails randmonly because of this double status update.
Besides, the test is wrong, as it does not expect the connector to be paused.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>