Fix invalid error message when topic RF change is disabled
When RF change is disabled, updating .spec.replicas on a KafkaTopic with 1 partiton and 1 replica fails with the following error:

Replication factor change not supported, but required for partitions
      []

The error message should contain the partition numbers (only 0 in this case), but we have an empty set.
This is because we use partitions instead of replicas when computing the partition set.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>