Deprecate reconciliationIntervalSeconds in favor of reconciliationIntervalMs
The spec field `reconciliationIntervalSeconds` is defined as int for the Topic Operator, and long for the User Operator. At runtime, this determines a different behavior when setting a value that overflows int. The User Operator will take and use it, while the Topic Operator will fail, which is better as it allows to quickly spot the configuration error.

Additionally, there is a mismatch between how the full reconciliation interval can be configured in the spec using `reconciliationIntervalSeconds`, where the time unit is seconds, and how the same can be done using the env var `STRIMZI_FULL_RECONCILIATION_INTERVAL_MS`, where the the time unit is milliseconds. This can easily lead to copy and paste errors.

This change deprecates `reconciliationIntervalSeconds` in favor of `reconciliationIntervalMs`, and uses long type for both operators. This aligns the configurations and matches the environment variable.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>