Fix TO-CC race condition issue happening during throttled rebalances
This change fixes the issue caused by the TO-CC race condition happening during throttled rebalances. The TO reverts dynamic throttled configs set by CC at the topic level because they don't match with the KT spec.

A private constant that contains the two throttle config properties is added to the BatchingTopiController, along with the following logic:

- If Cruise Control is enabled WITHOUT .spec.config throttle props, we skip any throttle alter config operation without warnings
- If Cruise Control is enabled WITH .spec.config throttle props, we do not skip the throttle alter config operations, but we raise a warning (can't be an error because there is no race when running rebalances without throttling, which is the default)
- If Cruise Control is NOT enabled, we reconcile as usual without warnings

The interaction between ALTERABLE_TOPIC_CONFIG and THROTTLING_CONFIG is covered by the new unit tests.

Fixes #9972.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>