Complete BatchingTopicController refactoring: handler classes
This change completes the BatchingTopicController refactoring by introduicing a set of handlers, one for each external system, that host the request logic.
With that, the BatchingTopicController class is only focused on the reconciliation logic, resulting in 40% less code and better maintenance.

Tested with all Topic Operator unit and integration tests, TopicST, and TopicReplicasChangeST.
I also ran the scalability test comparing this branch with 0.41.0, and the performance doesn't change.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>