Fix flaky TopicReplicasChangeST#testKafkaTopicReplicaChangePositiveRoundTrip
This change is an attempt to fix the flaky TopicReplicasChangeST#testKafkaTopicReplicaChangePositiveRoundTrip.
Looking at logs I can confirm that the timeout is indeed due to the Cruise Control's cluster model not being ready in time.

The Cruise Control setup proposed here seems to make the cluster model generation faster.
It basically reduces the partition/replica count of Cruise Control's topics and the metrics window.
On my machine, the whole test suite went down from 26 minutes to 13 minutes, which is a 50% improvement.
It would be good if someone else could confirm that by running TopicReplicasChangeST before and after this change.

With this change, I wasn't able to trigger the issue locally after many runs.
I would suggest to run regression tests 2 or 3 times to confirm this really helps.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>