Improve handling of invalid topic configurations
An invalid topic configuration results in a RuntimeException that fails the whole reconciliation and following periodic reconciliations.
Additionally, if you have a batch of 100 topics or more, it can be difficult to spot where the configuration errors is.

With this change the TO reports the error in the status of the Kafka topic containing the invalid configuration, and continues with the reconciliation.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>