Fix rebalance RF change race issue
When a rebalance is running, the Topic Operator reacts by seeing a RF change, where there is none, and raising the task already executing error.
This is caused by the admin.describeTopics returning no result for some time while the rebalance is in progress.
The Topic Operator translates this to a RF value of Integer.MIN_VALUE, which is different from any topic's RF.
This change fixes the issue by adding a RF greater than zero check.

Signed-off-by: Federico Valeri <fedevaleri@gmail.com>