Only roll for cluster CA key replacement in CaReconciler
Update CaReconciler to only roll pods when cluster CA
key is replaced, not when clients CA key is replaced.

Currently we do two rolling updates for the key replacement,
once in CaReconciler, and once in component reconcilers, e.g.
KafkaReconciler. This is not required for clients CA since
is is only used for trust.

Signed-off-by: Katherine Stanley <11195226+katheris@users.noreply.github.com>