Revert PR 10016 Allow KafkaRoller talk to controller directly
This reverts commit 1f1a16d2a57a13df990ef3fab9cc60386e4c0719.

This is because of the issue reported in KAFKA-18230,
admin requests for quorum info fail if the active controller has changed
and admin does not update its metadata and retry.
This causes errors in the operator, not able to reconcile Kafka nodes
until the metadata is eventually refreshed at some point.
This issue is fixed in Kafka 4.0.
Once only 4.0 and later versions are supported, this change will be re-implemented.

Signed-off-by: Gantigmaa Selenge <tina.selenge@gmail.com>